[CONTROLLER/GRPC] moves synchronizer service from trisolaris to grpc
