+++ server/controller/grpc/synchronizer/service.go
@@ -26,18 +26,19 @@ import (
 
 	"github.com/deepflowio/deepflow/server/controller/genesis"
 	grpcserver "github.com/deepflowio/deepflow/server/controller/grpc"
+	"github.com/deepflowio/deepflow/server/controller/grpc/statsd"
 	prometheus "github.com/deepflowio/deepflow/server/controller/prometheus/service/grpc"
-	"github.com/deepflowio/deepflow/server/controller/trisolaris/services/grpc/statsd"
+	trisolaris "github.com/deepflowio/deepflow/server/controller/trisolaris/services/grpc/synchronize"
 )
 
 type service struct {
-	vTapEvent                *VTapEvent
-	tsdbEvent                *TSDBEvent
-	ntpEvent                 *NTPEvent
-	upgradeEvent             *UpgradeEvent
-	kubernetesClusterIDEvent *KubernetesClusterIDEvent
-	processInfoEvent         *ProcessInfoEvent
-	pluginEvent              *PluginEvent
+	vTapEvent                *trisolaris.VTapEvent
+	tsdbEvent                *trisolaris.TSDBEvent
+	ntpEvent                 *trisolaris.NTPEvent
+	upgradeEvent             *trisolaris.UpgradeEvent
+	kubernetesClusterIDEvent *trisolaris.KubernetesClusterIDEvent
+	processInfoEvent         *trisolaris.ProcessInfoEvent
+	pluginEvent              *trisolaris.PluginEvent
 	prometheusEvent          *prometheus.SynchronizerEvent
 }
 
@@ -47,12 +48,12 @@ func init() {
 
 func newService() *service {
 	return &service{
-		vTapEvent:        NewVTapEvent(),
-		tsdbEvent:        NewTSDBEvent(),
-		ntpEvent:         NewNTPEvent(),
-		upgradeEvent:     NewUpgradeEvent(),
-		processInfoEvent: NewprocessInfoEvent(),
-		pluginEvent:      NewPluginEvent(),
+		vTapEvent:        trisolaris.NewVTapEvent(),
+		tsdbEvent:        trisolaris.NewTSDBEvent(),
+		ntpEvent:         trisolaris.NewNTPEvent(),
+		upgradeEvent:     trisolaris.NewUpgradeEvent(),
+		processInfoEvent: trisolaris.NewprocessInfoEvent(),
+		pluginEvent:      trisolaris.NewPluginEvent(),
 		prometheusEvent:  prometheus.NewSynchronizerEvent(),
 	}
 }
